+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 ### Added
 - Self hosted Bitbucket support
 ### Changed
+- Empty interface to any
 ### Deprecated
 ### Removed
 ### Fixed